+++ tests/test_jonga.py
@@ -1,5 +1,7 @@
 import pytest
 
+import os
+import tempfile
 import re
 import jonga
 
@@ -11,10 +13,27 @@ class TestSet01(object):
         ct.start()
         rec = re.compile('^[^\.]*.[^\.]*')
         ct.stop()
-        g = ct.graph()
-        assert(g is not None)
-        s = str(ct)
-        assert(s != '')
+        assert(ct.graph() is not None)
+        assert(str(ct) != '')
         ct.reset()
-        s = str(ct)
-        assert(s == '')
+        assert(str(ct) == '')
+
+
+    def test_02(self):
+        with jonga.ContextCallTracer(jonga.CallTracer(
+                                    grpflt='^[^\.]*')) as cct:
+            rec = re.compile('^[^\.]*.[^\.]*')
+        ct = cct.calltracer()
+        assert(str(ct) != '')
+        ct.reset()
+        assert(str(ct) == '')
+
+
+    def test_03(self):
+        ct = jonga.CallTracer(dstmodflt='^(re.|sre_)', grpflt='^[^\.]*')
+        fd, pth = tempfile.mkstemp(suffix='.svg')
+        os.close(fd)
+        with jonga.ContextCallTracer(ct, pth):
+            rec = re.compile('^[^\.]*.[^\.]*')
+        assert(os.path.getsize(pth) > 0)
+        os.remove(pth)